Added a nav bar to the secure subdomain
The application layout now renders the nav bar for the secure subdomain
if it doesn't recieve an override. I'd really much rather use a
layouts/secure.html.haml file and just override these things in
layouts/landing, but that doesn't seem to be an option (unless I create
a layout folder for every other scaffold that uses it which I might
do)... This solution works for now.

Minor optimization for the cars index page. Page loads in 4 queries
instead of 3N+1. On development this appears to be as fast or faster
than using a single query with left outer joins. Will analyze further on
production.

Added some indexes and limited the size of the Users table, should speed
things up very slightly. I'm going to do the same for Cars but really I
need a bigger sample size to test with for optimization to see how
indexes affect things

Also changed the user has_subscription? method to just return true for
the time being. I stupidly made it send a request to Stripe during
request processing on every page... Need to create a more robust system
using webhooks. (See #12)

Lots of code style changes, switched everything to ruby 1.9 style hash
syntax (I'm comfortable doing this now that we know rails 4 is going to
require 1.9 minimum and we're deployed on hardware that doesn't even
support versions earlier than 1.9.2)

Also changed a few pages where I saw a deprication warning on the
:confirm => :text method, changed to use the 4.0 syntax.

Changed the users model and associated form to allow email to be changed
by admin users. (I don't think there's a way to have attr_accessible as:
:admin to include all the attributes listed for the default role, will
look into it further.)

Closes #7