Finished styling car index page.
Defined the standard terms for our layout elements. In addition to the
navbar, container, and breadcrumbs, we have several boxes. The kind with
a header and content segment are just a box, those that are just content
are box-small. Small boxes have a gradient for their entire background
while normal boxes only have it in their header.

Finally found a work around to make the footer stick to the bottom of
the page.

These style changes will have broken other pages which need to be fixed