Changed the overal design of tow records display
Barely done. The info_lien_procedure tab is going to show the full
information for the current active tow_record, any others will be
relegated to the history tab. Get a scope on TowRecord that determines
if it's active and show that in the partial. Also need to write an
EFFICIENT method that determines if we have a history to show or not...